drapecol {shape}	R Documentation

draping colors over a persp plot

Description

generates color(s) that will appear on the surface facets of a "persp" plot.

Usage

drapecol(A, col=femmecol(100), NAcol="white")

Arguments

A 	 matrix with input grid.

col 	 color palette.

NAcol 	 color of NA elements.

Value

a vector of character strings giving the colors in hexadecimal format, one for each surface facet.

Note

This function is inspired by a similar function in package fields, unfortunately made unavailable in most recent version of fields

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>.

See Also

persp

Examples

  persp(volcano,theta = 135, phi = 30, col = drapecol(volcano),main="drapecol")
  persp(volcano,theta = 135, phi = 30, col = drapecol(volcano),border=NA,
        main="drapecol")

[Package shape version 1.2.2 Index]