plotcircle {shape}	R Documentation

adds part of a colored circle to a plot

Description

adds (part of) a colored circle to a plot; an arrow can be drawn at a specified position

Usage

plotcircle(r=1, ...)

Arguments

r 	 radius of circle.

... 	 arguments passed to function plotellipse.

Details

see plotellipse for details

Author(s)

Karline Soetaert <k.soetaert@nioo.knaw.nl>

See Also

plotellipse to draw ellipses

Examples

  emptyplot(c(0,1),c(0,2),main="plotcircle")
  plotcircle(mid=c(0.5,0.5),r=0.25,from=0,to=3*pi/2,
             arrow=TRUE,arr.pos=0.5,col="red")

[Package shape version 1.2.2 Index]