shape-package {shape}	R Documentation

Functions for plotting graphical shapes, colors

Description

Functions for plotting graphical shapes such as ellipses, circles, cylinders, arrows, ...

Support for the book "A practical guide to ecological modelling - using R as a simulation platform" by Karline Soetaert and Peter M.J. Herman (2009). Springer.

Details

Package: 	 shape

Type: 	 Package

Version: 	 1.2

Date: 	 2008-07-21

License: 	 GNU Public License 2 or above

This package is used in R-package ecolMod, which includes many more examples.

See also R-package diagram.

Author(s)

Karline Soetaert (Maintainer)

See Also

A4, writelabel, emptyplot, drapecol, femmecol, intpalette, shadepalette, colorlegend, greycol, rotatexy, Arrowhead, Arrows, cylindersegment, filledcylinder, filledcircle, filledellipse, filledmultigonal, filledrectangle, filledshape, getellipse, plotcircle, plotellipse, roundrect.

Examples

## Not run: 
## show examples (see respective help pages for details)
example(rotatexy)
example(filledshape)

## run demos
demo("colorshapes") # creating colored shapes

## open the directory with source code of demos
browseURL(paste(system.file(package="shape"), "/demo", sep=""))

## show package vignette
vignette("shape")
edit(vignette("shape"))
browseURL(paste(system.file(package="shape"), "/doc", sep=""))
## End(Not run)

[Package shape version 1.2.2 Index]